+++ store/create-cafs-store/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @pnpm/create-cafs-store
 
+## 6.0.0
+
+### Major Changes
+
+- 9caa33d53: `fromStore` replaced with `resolvedFrom`.
+
+### Patch Changes
+
+- Updated dependencies [9caa33d53]
+- Updated dependencies [9caa33d53]
+  - @pnpm/store-controller-types@17.0.0
+  - @pnpm/fs.indexed-pkg-importer@5.0.0
+  - @pnpm/store.cafs@2.0.4
+  - @pnpm/fetcher-base@15.0.1
+
 ## 5.1.1
 
 ### Patch Changes